chore: enable CodeQL scanning (#42)
CodeQL reviews a lot of things, among them, are the checks of
memory allocation and how to make sure to never do an overflow.

Closes #41

+++ controllers/cluster_controller.go
@@ -618,7 +618,7 @@ func (r *ClusterReconciler) ReconcilePods(ctx context.Context, cluster *apiv1.Cl
 	// cluster.Status.Instances == cluster.Spec.Instances and
 	// we don't need to modify the cluster topology
 	if cluster.Status.ReadyInstances != cluster.Status.Instances ||
-		cluster.Status.ReadyInstances != int32(len(instancesStatus.Items)) ||
+		cluster.Status.ReadyInstances != len(instancesStatus.Items) ||
 		!instancesStatus.IsComplete() {
 		contextLogger.Debug("Waiting for Pods to be ready")
 		return ctrl.Result{RequeueAfter: 1 * time.Second}, ErrNextLoop
